requirements.txt
asgiref==3.9.1
Brotli==1.1.0
click==8.2.1
dj-database-url==3.0.1
Django==5.2.5
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
environs==14.3.0
gunicorn==23.0.0
h11==0.16.0
marshmallow==4.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
uvicorn==0.35.0
whitenoise==6.9.0
